Convert all positions to doubles, from floats

+++ src/engine/FancyMath.java
@@ -1,5 +1,6 @@
 package engine;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import static engine.FancyMath.randomForceValue;
@@ -21,16 +22,15 @@ public class FancyMath {
         return randomNumber(x) * randomDirFloat();
     }
 
-    public static float getDistance(Vector3f pos1, Vector3f pos2){
+    public static float getDistance(Vector3d pos1, Vector3d pos2){
         return (float)Math.hypot((pos1.x - pos2.x), Math.hypot((pos1.y - pos2.y), (pos1.z - pos2.z)));
     }
 
-    //todo this does not belong in here
-    public static float getDistance(float x1, float y1, float z1, float x2, float y2, float z2){
-        float x = x1 - x2;
-        float y = y1 - y2;
-        float z = z1 - z2;
-        return (float)Math.hypot(x, Math.hypot(y,z));
+    public static double getDistance(double x1, double y1, double z1, double x2, double y2, double z2){
+        double x = x1 - x2;
+        double y = y1 - y2;
+        double z = z1 - z2;
+        return Math.hypot(x, Math.hypot(y,z));
     }
 
 }
